Hibakezeles

+++ haromszogek/haromszog.cs
@@ -5,9 +5,9 @@ namespace haromszogek
 {
     internal class haromszog
     {
-        private int aOldal;
-        private int bOldal;
-        private int cOldal;
+        private double aOldal;
+        private double bOldal;
+        private double cOldal;
 
         public double Terulet { get; private set; }
         public double Kerulet { get; private set; }
@@ -54,7 +54,7 @@ namespace haromszogek
             return aOldal + bOldal + cOldal;
 
         }
-        public haromszog(int aOldal, int bOldal, int cOldal)
+        public haromszog(double aOldal,double bOldal, double cOldal)
         {
             this.aOldal = aOldal;
             this.bOldal = bOldal;
